moving getTheme to hook
