+++ packages/docs/src/components/InfoPage.tsx
@@ -3,7 +3,7 @@ import { graphql } from 'gatsby';
 
 import Layout from './Layout';
 import { MdxQuery, TableOfContentsItem } from '../helpers/graphQLTypes';
-import { getTheme } from '../helpers/themeUtils';
+import useTheme from '../helpers/useTheme';
 import ContentRenderer from './ContentRenderer';
 
 /**
@@ -12,7 +12,7 @@ import ContentRenderer from './ContentRenderer';
 const InfoPage = ({ data, location }: MdxQuery) => {
   const { frontmatter, body, tableOfContents } = data.mdx;
   const { title, relatedUswdsGuidance, status } = frontmatter;
-  const theme = getTheme();
+  const theme = useTheme();
   let showGuidance = false;
 
   // check table of contents to see if there is a guidance section
