+++ packages/docs/src/components/ThemeSwitcher.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Dropdown } from '@cmsgov/design-system';
-import { getTheme } from '../helpers/themeUtils';
+import useTheme from '../helpers/useTheme';
 import { setQueryParam } from '../helpers/urlUtils';
 
 const themeOptions = [
@@ -13,6 +13,8 @@ const themeOptions = [
  * The dropdown component to switch the theme of the documentation site
  */
 const ThemeSwitcher = () => {
+  const currentTheme = useTheme();
+
   const onThemeChange = (e) => {
     // set the query param and reload the page for re-render
     setQueryParam('theme', e.target.value, true);
@@ -24,7 +26,7 @@ const ThemeSwitcher = () => {
       name="theme-switcher"
       options={themeOptions}
       className="c-theme-switcher"
-      value={getTheme()}
+      value={currentTheme}
       onChange={onThemeChange}
     />
   );
